+++ election/packages/ge2020/exitpolls/__tests__/empty.spec.js
@@ -1,0 +1,16 @@
+const { main } = require('../index')
+
+jest.resetModules()
+jest.mock('fs', () => ({
+  readFileSync: jest.fn(() => '{}'),
+}));
+
+describe('main method call', () => {
+  test('should return empty json response', () => {
+    // act
+    const output = main()
+
+    // assert
+    expect(output).toStrictEqual({ body: {} });
+  })
+})
