+++ election/packages/ge2020/exitpolls/__tests__/exists.spec.js
@@ -1,6 +1,6 @@
  
-const fs = require('fs');
+const fs = require('fs')
 
 test('should check existence of data file', () => {
-  expect(fs.existsSync('../../../transformed-data/exitpolls.json')).toBe(true);
+  expect(fs.existsSync('../../../transformed-data/exitpolls.json')).toBe(true)
 });
