+++ election/packages/ge2020/exitpolls/index.js
@@ -1,12 +1,9 @@
-const fs = require('fs')
+const fs = require('fs');
 
-let contents
-
+const contents = fs.readFileSync(__dirname + '/exitpolls.json', 'UTF-8');
+const body = JSON.parse(contents);
 function main(args) {
-	if (!contents) {
-		contents = fs.readFileSync(__dirname + '/exitpolls.json', 'UTF-8')
-	}
-	return { body: JSON.parse(contents) }
+    return { body };
 }
 
 module.exports = { main }
